supplier/requirements.txt
django-numpy==1.0.1